requirements.txt
agno
groq
python-dotenv
streamlit
coqui-tts
torch
soundfile
